# FilmApi/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "FilmApi",
    "version": "v1"
  },
  "paths": {
...
    "/Films/GetFilmGenre/{id}": {
      "get": {
        "tags": [
          "Films"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Films/GetConnectedGamesIds/{id}": {
      "get": {
        "tags": [
          "Films"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Films/GetConnectedBooksIds/{id}": {
      "get": {
        "tags": [
          "Films"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Films/GetFilmByGenre/{genres}": {
      "get": {
        "tags": [
          "Films"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "genres",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Films/getFilmByIdExternal/{id}": {
      "get": {
        "tags": [
          "Films"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Films/getFilmById/{id}": {
      "get": {
        "tags": [
          "Films"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Films/filmByName": {
      "get": {
        "tags": [
          "Films"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "filmName",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/Films/createFilm": {
      "post": {
        "tags": [
          "Films"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateFilmBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/HttpResponseMessage"
            }
          }
        }
      }
    },
    "/Films/deleteFilmById/{id}": {
      "delete": {
        "tags": [
          "Films"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Films/updateFilm/{id}": {
      "put": {
        "tags": [
          "Films"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FilmResponse"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Version": {
      "type": "object",
      "properties": {
        "major": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "minor": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "build": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "revision": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "majorRevision": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "minorRevision": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        }
      }
    },
    "StringStringIEnumerableKeyValuePair": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string",
          "readOnly": true
        },
        "value": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "readOnly": true
        }
      }
    },
    "HttpContent": {
      "type": "object",
      "properties": {
        "headers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StringStringIEnumerableKeyValuePair"
          },
          "readOnly": true
        }
      }
    },
    "HttpStatusCode": {
      "format": "int32",
      "enum": [
        100,
        101,
        102,
        103,
        200,
        201,
        202,
        203,
        204,
        205,
        206,
        207,
        208,
        226,
        300,
        301,
        302,
        303,
        304,
        305,
        306,
        307,
        308,
        400,
        401,
        402,
        403,
        404,
        405,
        406,
        407,
        408,
        409,
        410,
        411,
        412,
        413,
        414,
        415,
        416,
        417,
        421,
        422,
        423,
        424,
        426,
        428,
        429,
        431,
        451,
        500,
        501,
        502,
        503,
        504,
        505,
        506,
        507,
        508,
        510,
        511
      ],
      "type": "integer"
    },
    "HttpVersionPolicy": {
      "format": "int32",
      "enum": [
        0,
        1,
        2
      ],
      "type": "integer"
    },
    "HttpMethod": {
      "type": "object",
      "properties": {
        "method": {
          "type": "string",
          "readOnly": true
        }
      }
    },
    "HttpRequestMessage": {
      "type": "object",
      "properties": {
        "version": {
          "$ref": "#/definitions/Version"
        },
        "versionPolicy": {
          "$ref": "#/definitions/HttpVersionPolicy"
        },
        "content": {
          "$ref": "#/definitions/HttpContent"
        },
        "method": {
          "$ref": "#/definitions/HttpMethod"
        },
        "requestUri": {
          "format": "uri",
          "type": "string"
        },
        "headers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StringStringIEnumerableKeyValuePair"
          },
          "readOnly": true
        },
        "properties": {
          "type": "object",
          "additionalProperties": { },
          "readOnly": true
        },
        "options": {
          "type": "object",
          "additionalProperties": { },
          "readOnly": true
        }
      }
    },
    "HttpResponseMessage": {
      "type": "object",
      "properties": {
        "version": {
          "$ref": "#/definitions/Version"
        },
        "content": {
          "$ref": "#/definitions/HttpContent"
        },
        "statusCode": {
          "$ref": "#/definitions/HttpStatusCode"
        },
        "reasonPhrase": {
          "type": "string"
        },
        "headers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StringStringIEnumerableKeyValuePair"
          },
          "readOnly": true
        },
        "trailingHeaders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StringStringIEnumerableKeyValuePair"
          },
          "readOnly": true
        },
        "requestMessage": {
          "$ref": "#/definitions/HttpRequestMessage"
        },
        "isSuccessStatusCode": {
          "type": "boolean",
          "readOnly": true
        }
      }
    },
...
    "FilmResponse": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "original_title": {
          "type": "string"
        },
        "genre_ids": {
          "type": "string"
        },
        "overview": {
          "type": "string"
        },
        "release_date": {
          "type": "string"
        },
        "vote_average": {
          "format": "float",
          "type": "number"
        },
        "poster_path": {
          "type": "string"
        }
      }
    },
    "CreateFilmBody": {
      "type": "object",
      "properties": {
        "original_title": {
          "type": "string"
        },
        "genre_ids": {
          "type": "string"
        },
        "overview": {
          "type": "string"
        },
        "release_date": {
          "type": "string"
        },
        "vote_average": {
          "format": "float",
          "type": "number"
        },
        "poster_path": {
          "type": "string"
        }
      }
    },
...
  }
}